Updated DockArea.py and requirements.txt and added Helpsection
